Places fragment, timezone, remove old times

+++ app/src/main/java/com/ournet/weather/data/ILocation.java
@@ -5,7 +5,7 @@ package com.ournet.weather.data;
  */
 
 public interface ILocation {
-    float getLongitude();
-    float getLatitude();
+    double getLongitude();
+    double getLatitude();
     Integer getId();
 }
